==> notebooks/03_extract_expression_level.ipynb <==
import os
import sys
from pathlib import Path
from typing import List

import cv2
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
import skimage.io
from skimage.measure import regionprops_table
from skimage import exposure, measure

%load_ext autoreload
%autoreload 2

# Import path
module_path = str(Path.cwd().parents[0])
if module_path not in sys.path:
    sys.path.append(module_path)

from config import *

module_path = str(Path.cwd().parents[0] / "src")

if module_path not in sys.path:
    sys.path.append(module_path)

import utils as my_utils

# Define image and masks folder
info_csv = data_meta / "info_combined.csv"

df = pd.read_csv(info_csv)

df = df[df.After_bleach==False]

# Get nuclei, cyto, cell level features

def get_location(df, location):
    imgs_path = []
    markers = []
    df = df[df.Location == location]
    for row in df.itertuples():
        if row.Marker == 'Hoeschst' and row.Cycle != 'Cycle1':
            continue
        imgs_path.append(row.Path)
        markers.append(row.Marker)
    return imgs_path, markers


def get_region_props(img, cell, nuclei, cyto, marker, location, prop):
    """
    Get regionprops from masked region
    """
    cell_props = regionprops_table(cell, img, properties=prop)
    nuclei_props = regionprops_table(nuclei, img, properties=prop)
    cyto_props = regionprops_table(cyto, img, properties=prop)

    props = [cell_props, nuclei_props, cyto_props]
    prefix = ["Cell_", "Nuclei_", "Cyto_"]
    df_marker = pd.concat(
        [pd.DataFrame(prop).add_prefix(prefix[idx]) for idx, prop in enumerate(props)],
        axis=1,
    )
    df_marker["Marker"] = marker
    df_marker["Location"] = location
    df_marker["Id"] = regionprops_table(cell, properties=["label"])["label"]

    return df_marker


def measure_prop(df, masks, show=False, save=False):
    appended_data_intensity = []
    appended_data_morphology = []

    # Get each segmented cell region with corresponding nuclei and cell
    for idx, location in enumerate(masks.keys()):
        # Get cell masks
        mask_cyto = masks[location]["cyto"]
        mask_nuclei = masks[location]["nuclei"]
        cell, nuclei, cyto = my_utils.qc_nuclei(mask_cyto, mask_nuclei)
        
        # Get images
        imgs, markers = get_location(df, int(location))
        for idx, img_path in enumerate(imgs):
            img = skimage.io.imread(img_path)
            p2, p98 = np.percentile(img, (0, 99.9))
            img = exposure.rescale_intensity(img, in_range=(p2, p98))
#             print(img.shape, cell.shape, nuclei.shape, cyto.shape)
            df_marker = get_region_props(
                img, cell, nuclei, cyto, markers[idx], location, ["mean_intensity"]
            )
            appended_data_intensity.append(df_marker)

            # Get morphology only one time because morphology accross cell are the same
            if idx == 0:
                df_marker = get_region_props(
                    img, cell, nuclei, cyto, markers[idx], location, ["area"]
                )
                appended_data_morphology.append(df_marker)

                if show:
                    fig, ax = plt.subplots(1, 3, figsize=(15, 5))
                    ax[0].imshow(cell)
                    ax[1].imshow(nuclei)
                    ax[2].imshow(cyto)
                    plt.show()

    df_prop_intensity = pd.concat(appended_data_intensity, ignore_index=True)
    df_prop_intensity = df_prop_intensity.sort_values(
        by=["Location", "Id"]
    ).reset_index(drop=True)

    df_prop_morph = pd.concat(appended_data_morphology, ignore_index=True)
    df_prop_morph = df_prop_morph.sort_values(by=["Location", "Id"]).reset_index(
        drop=True
    )

    if save:
        df_prop_intensity.to_csv(data_meta / "measurements_intensity.csv", index=False)
        df_prop_morph.to_csv(data_meta / "measurements_morphology.csv", index=False)

    return df_prop_intensity, df_prop_morph

masks = my_utils.get_masks(data_mask)

df_prop_i, df_prop_m = measure_prop(df, masks, save=True, show=False)

df_prop_i

df_prop_m

# Get pixel level features

def get_location(df, location):
    imgs_path = []
    markers = []
    df = df[(df.Location == location)]
    for row in df.itertuples():
        imgs_path.append(row.Path)
        markers.append(row.Marker)
    return imgs_path, markers

def combined_imgs(imgs: List[str]) -> np.ndarray:
    """
    Return an multiplex image of dimension (num markers, height,width)
    """
    if not imgs:
        raise Exception("You passed a empty list of images path")
    pixels = []
    for img_path in imgs:
        img = skimage.io.imread(img_path)
        p2, p98 = np.percentile(img, (0, 99.9))
        img = exposure.rescale_intensity(img, in_range=(p2, p98), out_range=(0,1))
        pixels.append(img)
    return np.stack(pixels)

def non_border_cell(cell_mask):
    labels = []
    row_max, col_max = cell_mask.shape
    props = measure.regionprops(cell_mask)
    for prop in props:
        if prop['bbox'][0] == 0 or prop['bbox'][1] == 0 or prop['bbox'][2] == row_max or prop['bbox'][3] == col_max:
            continue
        else:
            labels.append(prop['label'])
    return labels

def get_multiplex_pix(df, masks, save=False):
    df_appended = []
    for idx, location in enumerate(masks.keys()):
        # Get pixels data
        imgs, markers = get_location(df, int(location))
        pixels = combined_imgs(imgs)
        
        # Get cell and nuclei mask
        mask_cyto = masks[location]["cyto"]
        mask_nuclei = masks[location]["nuclei"]
        cell, _, _ = my_utils.qc_nuclei(mask_cyto, mask_nuclei)
        
        labels = non_border_cell(cell)
        
        # Extracted all multiplex pixels in cell
        rows, cols = np.where(np.isin(cell, labels))
        cell_pixels = pixels[:, rows, cols]

        # Create dataframe
        df_pixels = pd.DataFrame(cell_pixels.T, columns=markers)
        df_pixels["Location"] = location
        df_pixels["X"] = rows
        df_pixels["Y"] = cols
        df_pixels["Id"] = cell[rows, cols]
        df_dapi = df_pixels.filter(like="Hoeschst")
        df_pixels = df_pixels.drop(df_dapi, axis=1)
        df_pixels.insert(
            0, column="Hoeschst", value=df_dapi.mean(axis=1).astype(np.uint16)
        )
        df_appended.append(df_pixels)

    df_pixels = pd.concat(df_appended, ignore_index=True)

    if save:
        df_pixels.to_csv(data_meta / "pixel_intensity.csv", index=False)

    return df_pixels

masks = my_utils.get_masks(data_mask)

df_pixels = get_multiplex_pix(df, masks, save=True)

df_pixels.describe()
